ls_17 reorganized app
