+++ store/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ProductsService } from './products/products.service';
+import { ProductsService } from './services/products.service';
 
 @NgModule({
   declarations: [
